feat: Complete application rename from Akhi_gpt to NafsAI
🎯 BREAKING CHANGE: Complete application rebranding

This commit implements a comprehensive rename of the entire application
from 'Akhi_gpt' to 'NafsAI' based on user feedback requesting proper
app naming.

## Package/Bundle Identifiers
- Updated pubspec.yaml name from 'akhi_gpt' to 'nafs_ai'
- Android: Changed namespace and applicationId to 'com.moarafat.nafs_ai'
- iOS: Updated all PRODUCT_BUNDLE_IDENTIFIER references
- macOS: Updated bundle identifier and product name
- Linux/Windows: Updated binary names and application IDs

## Display Names
- Android: App label changed to 'NafsAI'
- iOS: Updated CFBundleDisplayName and CFBundleName
- Web: Updated manifest.json and index.html
- All platforms now display 'NafsAI' to users

## Project Structure
- Moved Android MainActivity to new package structure
- Updated all platform-specific configuration files
- Updated binary names across all desktop platforms

## Code References
- Updated all HTTP headers from 'Akhi GPT' to 'NafsAI'
- Changed API referer URLs from 'akhi-gpt.app' to 'nafs-ai.app'
- Updated export filenames and user-facing text
- Updated welcome messages and onboarding text

## Documentation & Build
- Updated README.md with new repository references
- Added comprehensive CHANGELOG.md entry
- Updated .gitignore patterns for new naming
- Updated all documentation files

## Impact
- Users will see 'NafsAI' as the app name on all platforms
- App store listings will show the new name
- All technical references are consistently updated
- No remnants of old 'Akhi GPT' naming remain

Note: CI/CD workflow updates will be added in a separate commit
due to GitHub workflow scope permissions.

Closes: Application rebranding request
Affects: All platforms (Android, iOS, web, desktop)
Testing: Requires full platform testing before release
