+++ android/app/src/main/kotlin/com/moarafat/nafs_ai/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.moarafat.akhi_gpt
+package com.moarafat.nafs_ai
 
 import io.flutter.embedding.android.FlutterActivity
