+++ macos/Runner/Configs/AppInfo.xcconfig
@@ -5,10 +5,10 @@
 // 'flutter create' template.
 
 // The application's name. By default this is also the title of the Flutter window.
-PRODUCT_NAME = akhi_gpt
+PRODUCT_NAME = nafs_ai
 
 // The application's bundle identifier
-PRODUCT_BUNDLE_IDENTIFIER = com.moarafat.akhi_gpt
+PRODUCT_BUNDLE_IDENTIFIER = com.moarafat.nafs_ai
 
 // The copyright displayed in application information
 PRODUCT_COPYRIGHT = Copyright © 2025 com.moarafat. All rights reserved.
