+++ pubspec.yaml
@@ -1,5 +1,5 @@
-name: akhi_gpt
-description: "A new Flutter project."
+name: nafs_ai
+description: "NafsAI - Your supportive Muslim companion"
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
